GDMS: GCP-6256, GCP-6258, GCP-6259, GCP-6276, , GCP-6277, GCP-6278

+++ src/main/java/org/generationcp/middleware/dao/gdms/AccMetadataSetDAO.java
@@ -115,7 +115,7 @@ public class AccMetadataSetDAO extends GenericDAO<AccMetadataSet, Integer>{
         return new TreeSet<Integer>();
     }
     
-    public int countNIdsByMarkerIdsAndDatasetIdsAndNotGIds(List<Integer> datasetIds, List<Integer> markerIds, List<Integer> gIds)
+    public long countNIdsByMarkerIdsAndDatasetIdsAndNotGIds(List<Integer> datasetIds, List<Integer> markerIds, List<Integer> gIds)
             throws MiddlewareQueryException {
         try {
